requirements.txt
accelerate==1.9.0
afinn==0.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
autocuda==0.16
beautifulsoup4==4.13.4
benepar==0.2.0
bioc==2.1
blinker==1.9.0
blis==1.3.0
boostaug==2.3.5
boto3==1.39.14
botocore==1.39.14
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
conllu==4.5.3
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
datasets==4.0.0
decorator==5.2.1
Deprecated==1.2.18
dill==0.3.8
docopt==0.6.2
emoji==2.14.1
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.0
fastcoref==2.1.6
filelock==3.18.0
findfile==2.0.1
flair==0.15.1
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.2
grpcio-status==1.71.2
gspread==6.2.1
hf-xet==1.1.3
httplib2==0.22.0
huggingface-hub==0.34.3
hydra-core==1.3.2
idna==3.10
imodels==2.0.0
intervaltree==3.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lightning-utilities==0.14.3
locate==1.1.1
lxml==6.0.0
marisa-trie==1.2.1
maverick-coref==1.0.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
metric_visualizer==0.9.17
mlxtend==0.23.4
more-itertools==10.7.0
mplcursors==0.6
mpld3==0.5.10
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
seaborn
streamlit
multiprocess==0.70.16
murmurhash==1.0.13
narwhals==1.42.0
natsort==8.4.0
neleval==3.1.1
networkx==3.5
nltk==3.8.1
numpy==2.3.0
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.2.1
plotly==5.24.1
pptree==3.1
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf>=5.26.1,<6.0.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pysentimiento==0.7.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.2
pytorch-warmup==0.2.0
pytorch_revgrad==0.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.13.1
sacremoses==0.1.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segtok==1.5.11
seqeval==1.2.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spicy==0.16.0
sqlitedict==2.1.0
srsly==2.5.1
stack-data==0.6.3
stanza==1.10.1
streamlit==1.47.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
textblob==0.18.0.post0
thinc==8.3.6
threadpoolctl==3.6.0
tikzplotlib==0.10.1
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torch-struct==0.5
torchmetrics==1.7.4
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.2
transformers==4.54.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.4.0
vaderSentiment==3.3.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
Wikipedia-API==0.8.1
wordfreq==3.1.1
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.1
